agents/penny_assistant/backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-aiplatform
PyPDF2
python-dotenv
adk
adk-memorybank
adk-evaluation 